fix router views and componentes
