+++ src/router/Routers.js
@@ -15,9 +15,9 @@ import {
   Mypet,
 } from '../screens/index';
 import UploadPets from '../screens/Client/Upload/UploadPets';
-import PetInfo from '../screens/Client/Home/PetInfo/PetInfo';
+import PetInfo from '../screens/Client/PetInfo/PetInfo';
 import ChoiseUser from '../screens/Client/ChoiseUser/ChoiseUser';
-import ProfileScreen from '../screens/Client/Home/Proffile/ProfileScreen';
+import ProfileScreen from '../screens/Client/PetInfoScreen/PetInfoScreen';
 import Geolocation from '../screens/Client/Geolocalization/GeolocationUser';
 import WalkerView from '../screens/Client/WalkerView/WalkerView';
 import {styles} from './style';
